requirements.txt
psycopg[binary]==3.2.13
python-dotenv==1.0.1
pytest==8.3.5
